% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_mun_pop.R
\name{avg_mun_pop}
\alias{avg_mun_pop}
\title{Average municipality population over estimates}
\usage{
avg_mun_pop(sex)
}
\arguments{
\item{sex}{character. \code{male} or \code{female}.}
}
\value{
a tibble.
}
\description{
Average municipality population over estimates
}
