% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brpop-package.R
\docType{package}
\name{brpop-package}
\alias{brpop}
\alias{brpop-package}
\title{brpop: Brazilian Population Estimatives}
\description{
Datasets with yearly Brazilian population estimates from 'Brazilian Health Ministry' (DataSUS) and 'Universidade Federal do Rio Grande do Norte' (UFRN), aggregated by municipality, state, health region, sex, and age groups. The data in this package was manually downloaded from the DataSUS and UFRN websites and converted to tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rfsaldanha.github.io/brpop/}
}

}
\author{
\strong{Maintainer}: Raphael Saldanha \email{raphael.saldanha@icict.fiocruz.br} (\href{https://orcid.org/0000-0003-0652-8466}{ORCID})

}
\keyword{internal}
