% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus2024_mun_female_pop.R
\name{datasus2024_mun_female_pop}
\alias{datasus2024_mun_female_pop}
\title{Municipality yearly female population estimates per age group from 2000 to 2024}
\usage{
datasus2024_mun_female_pop()
}
\description{
A dataset containing female population estimates for Brazilian municipalities per age groups from 2000 to 2024.
}
\details{
The estimates were computed by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

\describe{
\item{code_muni}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
