% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_mun_female_pop.R
\docType{data}
\name{datasus_mun_female_pop}
\alias{datasus_mun_female_pop}
\title{Municipality yearly female population estimates per age group from 2000 to 2021}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2216016 rows and 4 columns.
}
\usage{
datasus_mun_female_pop
}
\description{
A dataset containing female population estimates for Brazilian municipalities per age groups from 2000 to 2021.
}
\details{
The estimates were computed by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

\describe{
\item{code_muni}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
\keyword{datasets}
