% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_mun_male_pop.R
\name{datasus_mun_male_pop}
\alias{datasus_mun_male_pop}
\title{Municipality yearly male population estimates per age group from 2000 to 2021}
\usage{
datasus_mun_male_pop()
}
\description{
A dataset containing male population estimates for Brazilian municipalities per age groups from 2000 to 2021.
}
\details{
The estimates were computed by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

\describe{
\item{code_muni}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
