% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_female_pop.R
\docType{data}
\name{mun_female_pop}
\alias{mun_female_pop}
\title{Municipality yearly female population estimates per age group}
\format{
A tibble with 2,216,016 rows and 4 variables:
\describe{
\item{mun}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
\source{
\url{http://tabnet.datasus.gov.br/cgi/deftohtm.exe?popsvs/cnv/popbr.def}
}
\usage{
mun_female_pop
}
\description{
A dataset containing female population estimates for Brazilian municipalities per age groups from 2000 to 2021.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.
}
\keyword{datasets}
