% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_female_pop_totals.R
\name{mun_female_pop_totals}
\alias{mun_female_pop_totals}
\title{Municipality yearly female population estimates totals}
\usage{
mun_female_pop_totals()
}
\arguments{
\item{source}{character. \code{bmh} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing female population estimates for Brazilian municipalities totals.
}
\seealso{
\link{bmh_mun_female_pop}, \link{ufrn_mun_female_pop}.
}
