% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_pop.R
\name{mun_pop}
\alias{mun_pop}
\title{Municipality yearly population estimates per age group}
\usage{
mun_pop(source = "datasus")
}
\arguments{
\item{source}{character. \code{datasus} for Brazilian Health Ministry estimates, \code{ufrn} for UFRN-DEM-LEPP estimates, \code{avg} for an average between both.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian municipalities per age groups.
}
\seealso{
\link{datasus_mun_male_pop}, \link{datasus_mun_female_pop}, \link{ufrn_mun_male_pop}, \link{ufrn_mun_female_pop}.
}
