% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_pop_age.R
\name{mun_pop_age}
\alias{mun_pop_age}
\title{Municipality yearly population estimates per age group}
\usage{
mun_pop_age(source = "datasus", sex = "all")
}
\arguments{
\item{source}{character. \code{datasus} for Brazilian Health Ministry old estimates (2000 to 2021), \code{datasus2024} for Brazilian Health Ministry new estimates (2000 to 2024), or \code{ufrn} for UFRN-DEM-LEPP estimates (2010 to 2030).}

\item{sex}{character. \code{all} for all population, \code{male} for male population, \code{female} for female population. Defaults for \code{all}.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian municipalities per age groups.
}
\seealso{
\link{datasus_mun_male_pop}, \link{datasus_mun_female_pop}, \link{ufrn_mun_male_pop}, \link{ufrn_mun_female_pop}.
}
