% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_reg_saude.R
\docType{data}
\name{mun_reg_saude}
\alias{mun_reg_saude}
\title{Municipality and health region code table}
\format{
A tibble with 5,597 rows and 2 variables:
\describe{
\item{code_muni}{municipality 6 digits code}
\item{cod_reg_saude}{health region 4 digits code}
}
}
\usage{
mun_reg_saude
}
\description{
A dataset containing municipality and health regions codes.
#'
}
\keyword{datasets}
