% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_reg_saude_449.R
\docType{data}
\name{mun_reg_saude_449}
\alias{mun_reg_saude_449}
\title{Municipality and health region (449) code table}
\format{
A tibble with 5,597 rows and 2 variables:
\describe{
\item{code_muni}{municipality 6 digits code}
\item{cod_reg_saude}{health region 5 digits code}
}
}
\usage{
mun_reg_saude_449
}
\description{
A dataset containing municipality and health regions codes, following the 449 units specification.
#'
}
\keyword{datasets}
