% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_sex_pop.R
\name{mun_sex_pop}
\alias{mun_sex_pop}
\title{Municipality yearly population estimates per age group and sex}
\usage{
mun_sex_pop()
}
\value{
A tibble.
}
\description{
This function binds UF male and female datasets (\code{mun_male_pop} and \code{mun_female_pop}) and includes a new variable called 'sex'.
}
\seealso{
\link{mun_male_pop}, \link{mun_female_pop}.
}
