% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsaude_pop.R
\name{regsaude_pop}
\alias{regsaude_pop}
\title{Health region yearly population estimates per age group}
\usage{
regsaude_pop(type = "standard", source = "bmh")
}
\arguments{
\item{type}{character. 'standard' or 'reg_saude_449'}

\item{source}{character. \code{bmh} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian health regions per age groups.
}
\seealso{
\link{regsaude_male_pop}, \link{regsaude_female_pop}.
}
