% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsaude_pop_age.R
\name{regsaude_pop_age}
\alias{regsaude_pop_age}
\title{Health region yearly population estimates per age group}
\usage{
regsaude_pop_age(type = "standard", source = "datasus", sex = "all")
}
\arguments{
\item{type}{character. 'standard' or 'reg_saude_449'}

\item{source}{character. \code{datasus} for Brazilian Health Ministry old estimates (2000 to 2021), \code{datasus2024} for Brazilian Health Ministry new estimates (2000 to 2024), or \code{ufrn} for UFRN-DEM-LEPP estimates (2010 to 2030).}

\item{sex}{character. \code{all} for all population, \code{male} for male population, \code{female} for female population. Defaults for \code{all}.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian health regions per age groups.
}
\seealso{
\link{regsaude_male_pop}, \link{regsaude_female_pop}.
}
