% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_female_pop.R
\docType{data}
\name{uf_female_pop}
\alias{uf_female_pop}
\title{UF yearly female population estimates per age group}
\format{
A tibble with 10,692 rows and 4 variables:
\describe{
\item{uf}{UF 2 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
\usage{
uf_female_pop
}
\description{
A dataset containing female population estimates for Brazilian UFs ("Unidades Federativas") per age groups from 2000 to 2021.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.
}
\keyword{datasets}
