% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_pop_age.R
\name{uf_pop_age}
\alias{uf_pop_age}
\title{UF yearly population estimates per age group}
\usage{
uf_pop_age(source = "datasus")
}
\arguments{
\item{source}{character. \code{datasus} for Brazilian Health Ministry old estimates (2000 to 2021), \code{datasus2024} for Brazilian Health Ministry new estimates (2000 to 2024), or \code{ufrn} for UFRN-DEM-LEPP estimates (2010 to 2030).}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian UFs ("Unidades Federativas") per age groups.
}
\seealso{
\link{datasus_mun_male_pop}, \link{datasus_mun_female_pop}, \link{ufrn_mun_male_pop}, \link{ufrn_mun_female_pop}.
}
