% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_pop_totals.R
\name{uf_pop_totals}
\alias{uf_pop_totals}
\title{UF yearly population estimates totals}
\usage{
uf_pop_totals(source = "bmh")
}
\arguments{
\item{source}{character. \code{bmh} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian UFs ("Unidades Federativas") totals.
}
\seealso{
\link{bmh_mun_male_pop}, \link{bmh_mun_female_pop}, \link{ufrn_mun_male_pop}, \link{ufrn_mun_female_pop}.
}
