% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_sex_pop.R
\name{uf_sex_pop}
\alias{uf_sex_pop}
\title{UF yearly population estimates per age group and sex}
\usage{
uf_sex_pop()
}
\value{
A tibble.
}
\description{
This function binds UF male and female datasets (\code{uf_male_pop} and \code{uf_female_pop}) and includes a new variable called 'sex'.
}
\seealso{
\link{uf_male_pop}, \link{uf_female_pop}.
}
