% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_sex_pop_totals.R
\name{uf_sex_pop_totals}
\alias{uf_sex_pop_totals}
\title{UF yearly population estimates per sex}
\usage{
uf_sex_pop_totals()
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian UFs per sex from 2000 to 2020.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

More details about the estimation methodology may be found here: \url{http://tabnet.datasus.gov.br/cgi/POPSVS/NT-POPULACAO-RESIDENTE-2000-2021.PDF}

\describe{
\item{coduf}{Municipality 6 digits code}
\item{year}{Year of the estimative}
\item{sex}{Sex}
\item{pop}{Population estimative}
}
}
