% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufrn_mun_female_pop.R
\docType{data}
\name{ufrn_mun_female_pop}
\alias{ufrn_mun_female_pop}
\title{Municipality yearly female population estimates per age group from UFRN-DEM-LEPP}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2339400 rows and 4 columns.
}
\usage{
ufrn_mun_female_pop
}
\description{
A dataset containing male population estimates for Brazilian municipalities per age groups from 2010 to 2030.
}
\details{
The estimates were computed by the LEPP laboratory from the Demography Department/UFRN.
}
\keyword{datasets}
