% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufrn_mun_male_pop.R
\docType{data}
\name{ufrn_mun_male_pop}
\alias{ufrn_mun_male_pop}
\title{Municipality yearly male population estimates per age group, from 2010 to 2030}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2339400 rows and 4 columns.
}
\usage{
ufrn_mun_male_pop
}
\description{
A dataset containing male population estimates for Brazilian municipalities per age groups from 2010 to 2030.
}
\details{
The estimates were computed by the LEPP laboratory from the Demography Department/UFRN.

\describe{
\item{code_muni}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
\references{
FREIRE, F.H.M.A; GONZAGA, M.R; QUEIROZ, B.L. Projeção populacional municipal com estimadores bayesianos, Brasil 2010 - 2030. In: Sawyer, D.O (coord.).  Seguridade Social Municipais. Projeto Brasil 3 Tempos. Secretaria  Especial de Assuntos Estratégicos da Presidência da República (SAE/SG/PR) , United Nations Development Programme, Brazil (UNDP) and International Policy  Centre for Inclusive Growth. Brasília (IPC-IG), 2019.
}
\keyword{datasets}
