% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brpop-package.R
\docType{package}
\name{brpop-package}
\alias{brpop}
\alias{brpop-package}
\title{brpop: Brazilian Population Estimatives}
\description{
Datasets with yearly (2000 to 2021) Brazilian population estimates from DataSUS/Brazilian Health Ministry, aggregated by state, municipality, sex, and age groups. The data in this package is manually downloaded from the DataSUS website and converted to tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rfsaldanha.github.io/brpop/}
}

}
\author{
\strong{Maintainer}: Raphael Saldanha \email{raphael.saldanha@icict.fiocruz.br} (\href{https://orcid.org/0000-0003-0652-8466}{ORCID})

}
\keyword{internal}
