% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brpop-package.R
\docType{package}
\name{brpop-package}
\alias{brpop}
\alias{brpop-package}
\title{brpop: Brazilian Population Estimatives}
\description{
Functions to handle and aggregate population estimates for Brazilian municipalities by sex and age groups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rfsaldanha.github.io/brpop/}
  \item Report bugs at \url{https://github.com/rfsaldanha/brpop/issues}
}

}
\author{
\strong{Maintainer}: Raphael Saldanha \email{raphael.de-freitas-saldanha@inria.fr} (\href{https://orcid.org/0000-0003-0652-8466}{ORCID})

}
\keyword{internal}
