% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_male_pop_totals.R
\name{mun_male_pop_totals}
\alias{mun_male_pop_totals}
\title{Municipality yearly male population estimates totals}
\usage{
mun_male_pop_totals()
}
\value{
A tibble.
}
\description{
This function provides a tibble containing male population estimates for Brazilian municipalities totals from 2000 to 2021.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

More details about the estimation methodology may be found here: \url{http://tabnet.datasus.gov.br/cgi/POPSVS/NT-POPULACAO-RESIDENTE-2000-2021.PDF}

\describe{
\item{mun}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{pop}{population estimative}
}
}
\seealso{
\link{mun_male_pop}.
}
