% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_male_pop_totals.R
\name{mun_male_pop_totals}
\alias{mun_male_pop_totals}
\title{Municipality yearly male population estimates totals}
\usage{
mun_male_pop_totals(source = "datasus")
}
\arguments{
\item{source}{character. \code{datasus} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing male population estimates for Brazilian municipalities totals.
}
\seealso{
\link{datasus_mun_male_pop}, \link{ufrn_mun_male_pop}.
}
