% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_pop.R
\name{mun_pop}
\alias{mun_pop}
\title{Municipality yearly population estimates per age group}
\format{
A tibble with 2,216,016 rows and 4 variables:
\describe{
\item{mun}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{age_group}{age group}
\item{pop}{population estimative}
}
}
\usage{
mun_pop()
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian municipalities per age groups from 2000 to 2021.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.
}
\seealso{
\link{mun_male_pop}, \link{mun_female_pop}.
}
