% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_pop_age.R
\name{mun_pop_age}
\alias{mun_pop_age}
\title{Municipality yearly population estimates per age group}
\format{
A tibble with the following variables.
\describe{
\item{codmun}{Municipality 6 digits code}
\item{year}{Year of the estimative}
\item{age_group}{Age group}
\item{pop}{Population estimative}
}
}
\usage{
mun_pop_age(age_group_option = "SVS2")
}
\arguments{
\item{age_group_option}{character or list. SVS1 or SVS2 for defaults age groups from DataSUS, or a list with two vectors, where the first vector are breaks, and the second vector are labels.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian municipalities per age group from 2000 to 2020.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

More details about the estimation methodology may be found here: \url{http://tabnet.datasus.gov.br/cgi/POPSVS/NT-POPULACAO-RESIDENTE-2000-2021.PDF}
}
