% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_sex_pop.R
\name{mun_sex_pop}
\alias{mun_sex_pop}
\title{Municipality yearly population estimates per age group and sex}
\usage{
mun_sex_pop(source = "datasus")
}
\arguments{
\item{source}{character. \code{datasus} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function binds municipality male and female estimates and includes a new variable called 'sex'.
}
\seealso{
\link{datasus_mun_male_pop}, \link{datasus_mun_female_pop}, \link{ufrn_mun_male_pop}, \link{ufrn_mun_female_pop}.
}
