% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_sex_pop_age.R
\name{mun_sex_pop_age}
\alias{mun_sex_pop_age}
\title{Municipality yearly population estimates per sex and age group}
\usage{
mun_sex_pop_age(age_group_option = "SVS2")
}
\arguments{
\item{age_group_option}{character or list. SVS1, SVS2 or a list with breaks and labels.

\describe{
\item{codmun}{municipality 6 digits code}
\item{year}{year of the estimative}
\item{Sex}{Sex}
\item{age_group}{age group}
\item{pop}{population estimative}
}}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian municipalities per sex and age groups from 2000 to 2020.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

More details about the estimation methodology may be found here: \url{http://tabnet.datasus.gov.br/cgi/POPSVS/NT-POPULACAO-RESIDENTE-2000-2021.PDF}
}
