% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsaude_female_pop.R
\name{regsaude_female_pop}
\alias{regsaude_female_pop}
\title{Health region yearly female population estimates per age group}
\usage{
regsaude_female_pop(type = "standard", source = "datasus")
}
\arguments{
\item{type}{character. 'standard' or 'reg_saude_449'}

\item{source}{character. \code{datasus} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing female population estimates for Brazilian health regions per age groups.
}
\seealso{
\link{datasus_mun_female_pop}, \link{ufrn_mun_female_pop}.
}
