% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsaude_male_pop.R
\name{regsaude_male_pop}
\alias{regsaude_male_pop}
\title{Health region yearly male population estimates per age group}
\usage{
regsaude_male_pop(type = "standard", source = "datasus")
}
\arguments{
\item{type}{character. 'standard' or 'reg_saude_449'}

\item{source}{character. \code{datasus} for Brazilian Health Ministry old estimates (2000 to 2021), \code{datasus2024} for Brazilian Health Ministry new estimates (2000 to 2024), or \code{ufrn} for UFRN-DEM-LEPP estimates (2010 to 2030).}
}
\value{
A tibble.
}
\description{
This function provides a tibble containing male population estimates for Brazilian health regions per age groups.
}
\seealso{
\link{datasus_mun_male_pop}, \link{ufrn_mun_male_pop}.
}
