% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsaude_pop_totals.R
\name{regsaude_pop_totals}
\alias{regsaude_pop_totals}
\title{Health regions yearly population estimates}
\usage{
regsaude_pop_totals()
}
\value{
A tibble.
}
\description{
This function provides a tibble containing population estimates for Brazilian health regions from 2000 to 2020.
}
\details{
The estimates were calculated by DataSUS (Brazilian Ministry of Health), manually downloaded from DataSUS website, and organized as a tibble.

More details about the estimation methodology may be found here: \url{http://tabnet.datasus.gov.br/cgi/POPSVS/NT-POPULACAO-RESIDENTE-2000-2021.PDF}

\describe{
\item{codregsaude}{Health regions code}
\item{year}{Year of the estimative}
\item{pop}{Population estimative}
}
}
