% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_sex_pop.R
\name{uf_sex_pop}
\alias{uf_sex_pop}
\title{UF yearly population estimates per age group and sex}
\usage{
uf_sex_pop(source = "bmh")
}
\arguments{
\item{source}{character. \code{bmh} for Brazilian Health Ministry estimates, or \code{ufrn} for UFRN-DEM-LEPP estimates.}
}
\value{
A tibble.
}
\description{
This function binds municipality male and female estimates and includes a new variable called 'sex'.
}
\seealso{
\link{bmh_mun_male_pop}, \link{bmh_mun_female_pop}, \link{ufrn_mun_male_pop}, \link{ufrn_mun_female_pop}.
}
